:recycle: refactor: 로또 클래스 내부에서 검증하도록 변경
