+++ src/controllers/Controller.js
@@ -3,6 +3,7 @@ import OutputView from '../views/OutputView.js';
 import validateMoney from '../models/validations/MoneyInputValidator.js';
 import validateWinning from '../models/validations/WinningInputValidator.js';
 import winningParser from '../utils/WinningParser.js';
+import Lotto from '../models/Lotto.js';
 
 class Controller {
   static async run() {
@@ -10,7 +11,7 @@ class Controller {
     validateMoney(money);
     OutputView.printLotto(money);
     const winning = await InputView.WinningInput();
-    validateWinning(winningParser(winning));
+    const winningLotto = new Lotto(winningParser(winning));
   }
 }
