+++ src/models/Lotto.js
@@ -1,3 +1,5 @@
+import validateWinning from './validations/WinningInputValidator.js';
+
 class Lotto {
   #numbers;
 
@@ -7,9 +9,11 @@ class Lotto {
   }
 
   #validate(numbers) {
-    if (numbers.length !== 6) {
-      throw new Error('[ERROR] 로또 번호는 6개여야 합니다.');
-    }
+    validateWinning(numbers);
+  }
+
+  getNumbers() {
+    return this.#numbers;
   }
 
   // TODO: 추가 기능 구현
