+++ src/utils/RandomLottery.js
@@ -1,5 +1,5 @@
 import { Random } from '@woowacourse/mission-utils';
-import Lotto from '../Lotto.js';
+import Lotto from '../models/Lotto.js';
 
 const drawLottery = () => {
   const draw = Random.pickUniqueNumbersInRange(1, 45, 6).sort((a, b) => a - b);
